pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "raddd"
version = "0.1.0"
description = "Reference RADDD stack"
authors = ["Nintorac"]

[tool.poetry.dependencies]
python = "^3.11"
pedalboard = "^0.8.7"
pandas = "^2.2.0"
ray = "^2.9.1"
dagster-dbt = "^0.22.3"
dagster-pandas = "^0.22.3"
dagster = "^1.6.3"
dagster-graphql = "^1.6.3"
dagster-webserver = "^1.6.3"
dbt-duckdb = "1.7.0"
dbt-core = "1.7.0"
mido = "^1.3.2"
pyarrow = "^15.0.0"
pydub = "^0.25.1"
